API GET Test adjusted
