+++ src/main/java/com/solvd/laba/carina/demo/api/user/GetUserMethods.java
@@ -1,21 +1,19 @@
-package com.solvd.laba.carina.demo.api;
+package com.solvd.laba.carina.demo.api.user;
 
 import com.zebrunner.carina.api.AbstractApiMethodV2;
 import com.zebrunner.carina.api.annotation.Endpoint;
-import com.zebrunner.carina.api.annotation.RequestTemplatePath;
 import com.zebrunner.carina.api.annotation.ResponseTemplatePath;
 import com.zebrunner.carina.api.annotation.SuccessfulHttpStatus;
 import com.zebrunner.carina.api.http.HttpMethodType;
 import com.zebrunner.carina.api.http.HttpResponseStatusType;
 import com.zebrunner.carina.utils.config.Configuration;
 
-@Endpoint(url = "${base_url}/users/1", methodType = HttpMethodType.DELETE)
-@RequestTemplatePath(path = "api/users/_delete/rq.json")
-@ResponseTemplatePath(path = "api/users/_delete/rs.json")
+@Endpoint(url = "${base_url}/users", methodType = HttpMethodType.GET)
+@ResponseTemplatePath(path = "api/users/_get/rs.json")
 @SuccessfulHttpStatus(status = HttpResponseStatusType.OK_200)
-public class DeleteUserMethod extends AbstractApiMethodV2 {
+public class GetUserMethods extends AbstractApiMethodV2 {
 
-    public DeleteUserMethod() {
+    public GetUserMethods() {
         replaceUrlPlaceholder("base_url", Configuration.getRequired("api_url"));
     }
 }
