--- src/test/java/com/solvd/laba/carina/demo/HelloWorld.java
@@ -1,27 +1,0 @@
-package com.solvd.laba.carina.demo;
-
-import java.lang.invoke.MethodHandles;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.testng.annotations.Test;
-
-import com.zebrunner.carina.core.IAbstractTest;
-import com.zebrunner.carina.core.registrar.ownership.MethodOwner;
-
-/**
- * This is Hello World carina-demo tests.
- *
- * @author qpsdemo
- */
-public class HelloWorld implements IAbstractTest {
-
-    private static final Logger LOGGER = LoggerFactory.getLogger(MethodHandles.lookup().lookupClass());
-
-    @Test()
-    @MethodOwner(owner = "qpsdemo")
-    public void helloWorld() {
-        LOGGER.info("Hello World!");
-    }
-
-}
