--- src/test/java/com/solvd/laba/carina/demo/MobileSampleTest.java
@@ -1,88 +1,0 @@
-package com.solvd.laba.carina.demo;
-
-import org.apache.commons.lang3.RandomStringUtils;
-import org.testng.Assert;
-import org.testng.annotations.Test;
-
-import com.zebrunner.carina.core.IAbstractTest;
-import com.solvd.laba.carina.demo.mobile.gui.pages.common.CarinaDescriptionPageBase;
-import com.solvd.laba.carina.demo.mobile.gui.pages.common.ContactUsPageBase;
-import com.solvd.laba.carina.demo.mobile.gui.pages.common.LoginPageBase;
-import com.solvd.laba.carina.demo.mobile.gui.pages.common.UIElementsPageBase;
-import com.solvd.laba.carina.demo.mobile.gui.pages.common.WebViewPageBase;
-import com.solvd.laba.carina.demo.mobile.gui.pages.common.WelcomePageBase;
-import com.solvd.laba.carina.demo.utils.MobileContextUtils;
-import com.solvd.laba.carina.demo.utils.MobileContextUtils.View;
-import com.zebrunner.agent.core.annotation.TestLabel;
-import com.zebrunner.carina.core.registrar.ownership.MethodOwner;
-import com.zebrunner.carina.utils.mobile.IMobileUtils;
-
-public class MobileSampleTest implements IAbstractTest, IMobileUtils {
-
-    @Test()
-    @MethodOwner(owner = "qpsdemo")
-    @TestLabel(name = "feature", value = {"mobile", "regression"})
-    public void testLoginUser() {
-        String username = "Test user";
-        String password = RandomStringUtils.randomAlphabetic(10);
-        WelcomePageBase welcomePage = initPage(getDriver(), WelcomePageBase.class);
-        Assert.assertTrue(welcomePage.isPageOpened(), "Welcome page isn't opened");
-        LoginPageBase loginPage = welcomePage.clickNextBtn();
-        Assert.assertFalse(loginPage.isLoginBtnActive(), "Login button is active when it should be disabled");
-        loginPage.typeName(username);
-        loginPage.typePassword(password);
-        loginPage.selectMaleSex();
-        loginPage.checkPrivacyPolicyCheckbox();
-        CarinaDescriptionPageBase carinaDescriptionPage = loginPage.clickLoginBtn();
-        Assert.assertTrue(carinaDescriptionPage.isPageOpened(), "Carina description page isn't opened");
-    }
-
-	@Test()
-    @MethodOwner(owner = "qpsdemo")
-    @TestLabel(name = "feature", value = {"mobile", "regression"})
-    public void testWebView() {
-        WelcomePageBase welcomePage = initPage(getDriver(), WelcomePageBase.class);
-        LoginPageBase loginPage = welcomePage.clickNextBtn();
-        loginPage.login();
-        WebViewPageBase webViewPageBase = initPage(getDriver(), WebViewPageBase.class);
-        MobileContextUtils contextHelper = new MobileContextUtils();
-
-        contextHelper.switchMobileContext(View.WEB_CARINA);
-        ContactUsPageBase contactUsPage = webViewPageBase.goToContactUsPage();
-
-        contextHelper.switchMobileContext(View.WEB_BROWSER, View.WEB_CARINA);
-
-        contactUsPage.typeName("John Doe");
-        contactUsPage.typeEmail("some@email.com");
-        contactUsPage.typeQuestion("This is a message");
-        contactUsPage.submit();
-        Assert.assertTrue(contactUsPage.isSuccessMessagePresent());
-    }
-
-    @Test()
-    @MethodOwner(owner = "qpsdemo")
-    @TestLabel(name = "feature", value = {"mobile", "acceptance"})
-    public void testUIElements() {
-        WelcomePageBase welcomePage = initPage(getDriver(), WelcomePageBase.class);
-        LoginPageBase loginPage = welcomePage.clickNextBtn();
-        CarinaDescriptionPageBase carinaDescriptionPage = loginPage.login();
-        UIElementsPageBase uiElements = carinaDescriptionPage.navigateToUIElementsPage();
-        final String text = "some text";
-        final String date = "22/10/2018";
-        final String email = "some@email.com";
-        uiElements.typeText(text);
-        Assert.assertEquals(uiElements.getText(), text, "Text was not typed");
-        uiElements.typeDate(date);
-        Assert.assertEquals(uiElements.getDate(), date, "Date was not typed");
-        uiElements.typeEmail(email);
-        Assert.assertEquals(uiElements.getEmail(), email, "Email was not typed");
-        uiElements.swipeToFemaleRadioButton();
-        uiElements.checkCopy();
-        Assert.assertTrue(uiElements.isCopyChecked(), "Copy checkbox was not checked");
-        uiElements.clickOnFemaleRadioButton();
-        Assert.assertTrue(uiElements.isFemaleRadioButtonSelected(), "Female radio button was not selected!");
-        uiElements.clickOnOtherRadioButton();
-        Assert.assertTrue(uiElements.isOthersRadioButtonSelected(), "Others radio button was not selected!");
-    }
-
-}
