--- src/test/resources/api/soap/addinteger/rq.xml
@@ -1,9 +1,0 @@
-<soapenv:Envelope xmlns:soapenv="http://schemas.xmlsoap.org/soap/envelope/" xmlns:tem="http://tempuri.org">
-    <soapenv:Header/>
-    <soapenv:Body>
-        <tem:AddInteger>
-            <tem:Arg1>${firstNumber}</tem:Arg1>
-            <tem:Arg2>${secondNumber}</tem:Arg2>
-        </tem:AddInteger>
-    </soapenv:Body>
-</soapenv:Envelope>
