--- src/test/resources/api/soap/addinteger/rs.xml
@@ -1,9 +1,0 @@
-<SOAP-ENV:Envelope xmlns:SOAP-ENV="http://schemas.xmlsoap.org/soap/envelope/">
-    <SOAP-ENV:Body>
-        <AddIntegerResponse xmlns="http://tempuri.org">
-            <#if firstNumber?? and secondNumber??>
-                <AddIntegerResult>${firstNumber+secondNumber}</AddIntegerResult>
-            </#if>
-        </AddIntegerResponse>
-    </SOAP-ENV:Body>
-</SOAP-ENV:Envelope>
