--- src/test/resources/api/soap/soap.properties
@@ -1,9 +1,0 @@
-#=============== AddInteger properties ===========#
-firstNumber=4
-secondNumber=6
-result=10
-
-#=============== LookupCity properties ===========#
-zipCode=10005
-city=New York
-state=NY
