--- src/test/resources/data_source/calculator.csv
@@ -1,4 +1,0 @@
-Execute,TUID,a,b,c
-Y,CSV0001,2,6,8
-Y,CSV0002,1,12,13
-Y,CSV0003,99,1,100
