--- src/test/resources/data_source/finances.csv
@@ -1,9 +1,0 @@
-Execute,CompanyID,income,expenses,country
-Y,DGD01,1000,200,ES
-Y,DSA23,4322,3123,FR
-Y,KLR50,2133,222,DE
-Y,ARF21,370,342,IT
-Y,KSA95,2333,2800,IT
-Y,QWE90,7133,8222,FR
-Y,DSI56,9532,7299,FR
-Y,RKP80,566,222,DE
