--- src/test/resources/data_source/financesPrevYear.csv
@@ -1,9 +1,0 @@
-To test,id,income,expenses,country
-+,DGD01,800,400,ES
-+,DSA23,3422,3123,FR
-+,KLR50,1720,922,DE
-+,ARF21,770,342,IT
-+,KSA95,2933,2800,IT
-+,QWE90,7133,8222,FR
-+,DSI56,9532,7299,FR
-+,RKP80,566,222,DE
