--- src/test/resources/testng_suites/autodownload.xml
@@ -1,20 +1,0 @@
-<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd">
-<suite verbose="1" name="Carina Demo Tests - Auto Download" parallel="methods" thread-count="3">
-
-    <parameter name="suiteOwner" value="qpsdemo"/>
-    <parameter name="jenkinsJobName" value="Auto-Download-Test"/>
-    <parameter name="jenkinsJobType" value="web"/>
-    <parameter name="jenkinsEnvironments" value="DEMO"/>
-    <parameter name="jenkinsEmail" value="demo@zebrunner.com"/>
-
-    <parameter name="jenkinsRegressionPipeline" value="web_regression, full_regression"/>
-
-    <parameter name="jenkinsRegressionMatrix_full_regression" value="browser:chrome; browser:firefox"/>
-
-    <test name="Auto download test">
-        <classes>
-            <class name="com.solvd.laba.carina.demo.AutoDownloadTest"/>
-        </classes>
-    </test>
-
-</suite>
