--- src/test/resources/testng_suites/dataprovider.xml
@@ -1,82 +1,0 @@
-<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd">
-<suite verbose="1" name="Carina Demo - Test Custom Jobs Params" parallel="tests" data-provider-thread-count="30" thread-count="5">
-
-	<parameter name="suiteOwner" value="qpsdemo"/>
-	<parameter name="jenkinsJobName" value="API-CustomParams-Demo"/>
-	<parameter name="jenkinsJobType" value="api"/>
-	<parameter name="jenkinsEnvironments" value="DEMO"/>
-	<parameter name="jenkinsEmail" value="demo@zebrunner.com"/>
-
-	<parameter name="jenkinsRegressionPipeline" value="api_regression, full_regression"/>
-	<parameter name="jenkinsJobExecutionMode" value="continue"/>
-
-	<!-- example declaring non default retry_count -->
-	<parameter name="jenkinsDefaultRetryCount" value="2"/>
-
-	<!-- declaring custom arguments-->
-	<parameter name="stringParam::stringArg::custom string desc" value="string value" />
-	<parameter name="booleanParam::booleanArg::custom boolean desc" value="true" />
-	<parameter name="hiddenParam::hiddenArg::custom hidden desc" value="hidden value" />
-	<parameter name="choiceParam::choiceArg::custom choice desc" value="1, 2, 3" />
-
-	<test name="TestNG data providers test" parallel="methods">
-		<parameter name="a" value="12"/>
-		<parameter name="b" value="3"/>
-		<parameter name="c" value="9"/>
-		<classes>
-			<class name="com.solvd.laba.carina.demo.DataProvidersSampleTest">
-				<methods>
-					<include name="testSubtractOperation"/>
-					<include name="testMultiplyOperation"/>
-				</methods>
-			</class>
-		</classes>
-	</test>
-
-	<test name="Sum operation test" parallel="methods">
-		<classes>
-			<class name="com.solvd.laba.carina.demo.CustomDataProvidersSampleTest">
-				<methods>
-					<include name="testXlsSumOperation"/>
-					<include name="testCsvSumOperation"/>
-					<include name="testSumOperationFromMap"/>
-				</methods>
-			</class>
-		</classes>
-	</test>
-
-	<test name="Group by country operation test" parallel="methods">
-		<classes>
-			<class name="com.solvd.laba.carina.demo.CustomDataProvidersSampleTest">
-				<methods>
-					<include name="testDataSourceGroup"/>
-				</methods>
-			</class>
-		</classes>
-	</test>
-
-	<test name="Data provider static parameter test" parallel="methods">
-		<parameter name="expectedValue" value="arg"/>
-        <classes>
-            <class name="com.solvd.laba.carina.demo.CustomDataProvidersSampleTest">
-                <methods>
-                    <include name="staticParamTest"/>
-                </methods>
-            </class>
-        </classes>
-	</test>
-
-	<test name="Group by country with params from suite test" parallel="methods">
-		<parameter name="{ds_file}" value="data_source/financesPrevYear.csv"/>
-		<parameter name="{ds_uid}" value="id"/>
-		<parameter name="{ds_execute_column}" value="To test"/>
-		<parameter name="{ds_execute_value}" value="+"/>
-		<classes>
-			<class name="com.solvd.laba.carina.demo.CustomDataProvidersSampleTest">
-				<methods>
-					<include name="testDataSourceGroup"/>
-				</methods>
-			</class>
-		</classes>
-	</test>
-</suite>
