--- src/test/resources/testng_suites/helloWorld.xml
@@ -1,16 +1,0 @@
-<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd">
-<suite verbose="1" name="Carina Demo Tests - Hello World" parallel="methods">
-
-	<parameter name="suiteOwner" value="qpsdemo"/>
-	<parameter name="jenkinsJobName" value="Hello-World-Test"/>
-	<parameter name="jenkinsJobType" value="api"/>
-
-	<parameter name="jenkinsEnvironments" value="DEMO"/>
-
-	<test name="Hello World test">
-		<classes>
-			<class name="com.solvd.laba.carina.demo.HelloWorld" />
-		</classes>
-	</test>
-
-</suite>
