--- src/test/resources/testng_suites/ios-preferences.xml
@@ -1,24 +1,0 @@
-<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd">
-<suite name="Carina Demo Tests - iOS Preferences Sample" verbose="1" parallel="tests">
-
-	<parameter name="suiteOwner" value="qpsdemo" />
-	<parameter name="provider" value="mcloud"/>
-	<parameter name="jenkinsJobName" value="Mobile-iOS-Preferences-Test" />
-	<parameter name="jenkinsJobType" value="ios"/>
-	<parameter name="jenkinsEnvironments" value="DEMO" />
-	<parameter name="capabilities" value="bundleId=com.apple.Preferences;platformName=iOS;deviceName=ANY" />
-
-	<parameter name="jenkinsMobileDefaultPool" value="ANY" />
-
-	<parameter name="jenkinsRegressionPipeline" value="mobile_regression" />
-	<parameter name="jenkinsJobExecutionOrder" value="4" />
-	<parameter name="jenkinsJobExecutionMode" value="continue" />
-	<parameter name="jenkinsAutoScreenshot" value="true"/>
-
-	<test name="Preferences test">
-		<classes>
-			<class name="com.solvd.laba.carina.demo.IOSPreferencesTest" />
-		</classes>
-	</test>
-
-</suite>
