--- src/test/resources/testng_suites/ios-safari.xml
@@ -1,24 +1,0 @@
-<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd">
-<suite name="Carina Demo Tests - iOS Safari Sample" verbose="1" parallel="tests">
-
-	<parameter name="suiteOwner" value="qpsdemo" />
-	<parameter name="provider" value="mcloud"/>
-	<parameter name="jenkinsJobName" value="Mobile-iOS-Safari-Test" />
-	<parameter name="jenkinsJobType" value="web"/>
-	<parameter name="capabilities" value="browserName=safari;platformName=iOS;deviceName=ANY" />
-	<parameter name="jenkinsEnvironments" value="DEMO" />
-
-	<parameter name="jenkinsMobileDefaultPool" value="ANY" />
-
-	<parameter name="jenkinsRegressionPipeline" value="mobile_regression" />
-	<parameter name="jenkinsJobExecutionOrder" value="5" />
-	<parameter name="jenkinsJobExecutionMode" value="continue" />
-	<parameter name="jenkinsAutoScreenshot" value="true"/>
-
-	<test name="Safari test">
-		<classes>
-			<class name="com.solvd.laba.carina.demo.IOSSafariTest" />
-		</classes>
-	</test>
-
-</suite>
