--- src/test/resources/testng_suites/ios.xml
@@ -1,26 +1,0 @@
-<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd">
-<suite name="Carina Demo Tests - Mobile Sample (iOS)" verbose="1" parallel="tests">
-
-	<parameter name="suiteOwner" value="qpsdemo" />
-	<parameter name="provider" value="mcloud"/>
-	<parameter name="jenkinsJobName" value="Mobile-iOS-Simulator-Demo-Test" />
-	<parameter name="jenkinsJobType" value="ios"/>
-	<parameter name="jenkinsEmail" value="demo@zebrunner.com" />
-	<parameter name="jenkinsEnvironments" value="DEMO" />
-
-	<parameter name="stringParam::capabilities.app::path to carina demo mobile application" value="https://qaprosoft.s3-us-west-2.amazonaws.com/carinademoexample.zip" />
-
-	<parameter name="jenkinsMobileDefaultPool" value="ANY" />
-
-	<parameter name="jenkinsRegressionPipeline" value="mobile_regression" />
-	<parameter name="jenkinsJobExecutionOrder" value="3" />
-	<parameter name="jenkinsJobExecutionMode" value="continue" />
-	<parameter name="jenkinsAutoScreenshot" value="true"/>
-
-	<test name="Mobile test">
-		<classes>
-			<class name="com.solvd.laba.carina.demo.MobileSampleTest" />
-		</classes>
-	</test>
-
-</suite>
