--- src/test/resources/testng_suites/soap.xml
@@ -1,18 +1,0 @@
-<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd">
-<suite verbose="1" name="Carina Demo Tests - Soap Sample" parallel="methods" thread-count="5">
-
-	<parameter name="suiteOwner" value="qpsdemo"/>
-	<parameter name="jenkinsJobName" value="Soap-Demo-Test"/>
-	<parameter name="jenkinsJobType" value="api"/>
-
-	<parameter name="jenkinsEnvironments" value="DEMO"/>
-
-	<parameter name="jenkinsRegressionPipeline" value="api_regression, full_regression"/>
-
-	<test name="Soap API tests">
-		<classes>
-			<class name="com.solvd.laba.carina.demo.SoapSampleTest" />
-		</classes>
-	</test>
-
-</suite>
