--- src/test/resources/testng_suites/web-localization.xml
@@ -1,29 +1,0 @@
-<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd">
-<suite verbose="1" name="Carina Demo Tests - Web Localization Sample" parallel="methods">
-
-    <parameter name="suiteOwner" value="qpsdemo"/>
-    <parameter name="jenkinsJobName" value="Web-Localization-Demo-Test"/>
-    <parameter name="jenkinsJobType" value="web"/>
-    <parameter name="jenkinsEnvironments" value="DEMO"/>
-    <parameter name="jenkinsEmail" value="demo@zebrunner.com"/>
-
-    <parameter name="jenkinsRegressionPipeline" value="web_regression, full_regression"/>
-
-    <parameter name="jenkinsRegressionMatrix_full_regression" value="browser:chrome, locale:en_US; browser:chrome, locale:es_ES; browser:chrome, locale:de_DE; browser:chrome, locale:zh_CN; browser:chrome, locale:ru_RU; browser:chrome, locale:hu_HU; browser:chrome, locale:ja_JP"/>
-    <parameter name="jenkinsJobExecutionOrder" value="1"/>
-    <parameter name="jenkinsJobExecutionMode" value="continue"/>
-
-    <parameter name="choiceParam::locale::supported locales in example" value="en_US,es_ES,de_DE,zh_CN,ru_RU,hu_HU,ja_JP" />
-    <parameter name="jenkinsAutoScreenshot" value="true"/>
-
-    <test name="Wikipedia Localization Test">
-        <classes>
-            <class name="com.solvd.laba.carina.demo.WebLocalizationSample" >
-                <methods>
-                    <include name="testLanguages" />
-                </methods>
-            </class>
-        </classes>
-    </test>
-
-</suite>
