--- src/test/resources/testng_suites/web-multiple-browser.xml
@@ -1,19 +1,0 @@
-<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd">
-<suite verbose="1" name="Carina Demo Tests - Web Multiple Browser Sample" parallel="methods">
-
-    <parameter name="suiteOwner" value="qpsdemo"/>
-    <parameter name="jenkinsJobName" value="Web-Multiple-Browser-Demo-Test"/>
-    <parameter name="jenkinsJobType" value="web"/>
-    <parameter name="jenkinsEnvironments" value="DEMO"/>
-    <parameter name="jenkinsEmail" value="demo@zebrunner.com"/>
-
-    <parameter name="jenkinsRegressionPipeline" value="web_regression"/>
-    <parameter name="jenkinsAutoScreenshot" value="true"/>
-
-    <test name="Multiple Browser Test">
-        <classes>
-            <class name="com.solvd.laba.carina.demo.WebMultipleBrowserTest"/>
-        </classes>
-    </test>
-
-</suite>
