--- src/test/resources/testng_suites/web-open-strategy.xml
@@ -1,23 +1,0 @@
-<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd">
-<suite verbose="1" name="Carina Demo Tests - Web open stategy sample" parallel="methods">
-
-    <parameter name="suiteOwner" value="qpsdemo"/>
-    <parameter name="jenkinsJobName" value="Web-OpenStrategy-Test"/>
-    <parameter name="jenkinsJobType" value="web"/>
-    <parameter name="jenkinsEnvironments" value="DEMO"/>
-    <parameter name="jenkinsEmail" value="demo@zebrunner.com"/>
-
-    <parameter name="jenkinsRegressionPipeline" value="web_regression, full_regression"/>
-
-    <parameter name="jenkinsRegressionMatrix_full_regression" value="browser:chrome; browser:firefox"/>
-    <parameter name="jenkinsJobExecutionOrder" value="1"/>
-    <parameter name="jenkinsJobExecutionMode" value="continue"/>
-    <parameter name="jenkinsAutoScreenshot" value="true"/>
-
-    <test name="Gsm arena page open strategy test">
-        <classes>
-            <class name="com.solvd.laba.carina.demo.PageOpeningStrategySampleTest"/>
-        </classes>
-    </test>
-
-</suite>
